feat: Lv4. 로그인(인증) 세션을 활용해 로그인 기능 구현("/login") 필터를 활용해 공통된 인증처리(로그인, 회원가입은 제외) 로그인은 이메일과 비밀번호를 사용하며 일정의 작성, 수정, 삭제에는 세션값을 활용하여 유저를 판별(인가)

+++ src/main/java/com/example/jpascheduleapi/schedule/service/ScheduleService.java
@@ -1,25 +1,37 @@
 package com.example.jpascheduleapi.schedule.service;
 
+import com.example.jpascheduleapi.schedule.dto.ScheduleAuthorDto;
 import com.example.jpascheduleapi.schedule.dto.ScheduleRequestDto;
 import com.example.jpascheduleapi.schedule.dto.ScheduleResponseDto;
 import com.example.jpascheduleapi.schedule.entity.Schedule;
 import com.example.jpascheduleapi.schedule.repository.ScheduleRepository;
+import com.example.jpascheduleapi.user.dto.UserResponseDto;
+import com.example.jpascheduleapi.user.entity.User;
+import com.example.jpascheduleapi.user.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @RequiredArgsConstructor
 @Service
 public class ScheduleService {
     private final ScheduleRepository scheduleRepository;
+    private final UserRepository userRepository;
 
-    public ScheduleResponseDto createSchedule(ScheduleRequestDto requestDto) {
-        Schedule schedule = new Schedule(requestDto.getUsername(), requestDto.getTitle(), requestDto.getContents());
+    public ScheduleResponseDto createSchedule(ScheduleRequestDto requestDto, UserResponseDto loginUser) {
+        User user = userRepository.findUserByIdOrElseThrow(loginUser.getId());
+
+        Schedule schedule = new Schedule(requestDto.getTitle(), requestDto.getContents());
+        schedule.setUser(user);
+
         Schedule savedSchedule = scheduleRepository.save(schedule);
 
         return ScheduleResponseDto.builder()
                 .id(savedSchedule.getId())
-                .username(savedSchedule.getUsername())
+                // SchudeleAuthorDto는 String username 필드를 가지고 있음.
+                .user(new ScheduleAuthorDto(savedSchedule.getUser().getUsername()))
                 .title(savedSchedule.getTitle())
                 .contents(savedSchedule.getContents())
                 .createAt(savedSchedule.getCreatedAt())
@@ -28,9 +40,10 @@ public class ScheduleService {
 
     public ScheduleResponseDto findScheduleById(Long id) {
         Schedule foundSchedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
+        User user = foundSchedule.getUser();
 
         return ScheduleResponseDto.builder()
-                .username(foundSchedule.getUsername())
+                .user(new ScheduleAuthorDto(user.getUsername()))
                 .title(foundSchedule.getTitle())
                 .contents(foundSchedule.getContents())
                 .createAt(foundSchedule.getCreatedAt())
@@ -39,22 +52,31 @@ public class ScheduleService {
     }
 
     @Transactional
-    public ScheduleResponseDto updateSchedule(Long id, ScheduleRequestDto requestDto) {
+    public ScheduleResponseDto updateSchedule(Long id, ScheduleRequestDto requestDto, UserResponseDto loginUser) {
         Schedule foundSchedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
+
+        //todo 이 검증도 반복된다면 필터 사용 가능한지 확인
+        if (!foundSchedule.getUser().getId().equals(loginUser.getId())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "수정은 작성자만 가능합니다.");
+        }
 
         foundSchedule.updateTitle(requestDto.getTitle());
         foundSchedule.updateContents(requestDto.getContents());
 
         return ScheduleResponseDto.builder()
-                .username(foundSchedule.getUsername())
                 .title(foundSchedule.getTitle())
                 .contents(foundSchedule.getContents())
                 .modifiedAt(foundSchedule.getModifiedAt())
                 .build();
     }
 
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long id, UserResponseDto loginUser) {
         Schedule foundSchedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
+
+        if (!foundSchedule.getUser().getId().equals(loginUser.getId())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "삭제는 작성자만 가능합니다.");
+        }
+
         scheduleRepository.delete(foundSchedule);
     }
 }
